refactor: Extract methods in files from konnectorAlerts
